+++ utils/build_rego_file.py
@@ -1,4 +1,3 @@
-from typing import Dict, List
 from .commands_map import commands_map
