+++ utils/write_rego.py
@@ -1,7 +1,7 @@
-from rego_build_api.server.models import RequestObject
+from schemas.rules import RequestObject
 from .build_rego_file import build_rego
-from rego_build_api.server.github import initialize_repo, git_push
-from ..config.config import settings
+from server.github import initialize_repo, git_push
+from config.config import settings
 
 initiate_rule = "package httpapi.authz\nimport input\ndefault allow = false\n\n\n\n"
 
@@ -14,8 +14,10 @@ def write_to_file(rule: RequestObject) -> dict:
     """
 
     # Initialize repository
-    initialization_response = initialize_repo(settings.GITHUB_URL, settings.GITHUB_EMAIL, settings.GITHUB_USERNAME)
+    initialization_response = initialize_repo(settings.GITHUB_URL)
+
     repo_path = initialization_response["repo_path"]
+    repo_git_path = initialization_response["repo_git_path"]
 
     # Create rego file.
     with open(f"{repo_path}/auth.rego", "w") as file:
@@ -24,6 +26,6 @@ def write_to_file(rule: RequestObject) -> dict:
         file.write(result)
 
     # Push to GitHub
-    git_push(repo_path)
+    git_push(repo_path, repo_git_path)
 
     return {"status": "success"}
